+++ oh-my-just-open/oh_my_just_openApp.swift
@@ -2,10 +2,17 @@ import SwiftUI
 
 @main
 struct oh_my_just_openApp: App {
+    @State private var updateService = UpdateService()
+
     var body: some Scene {
         WindowGroup {
             ContentView()
         }
         .windowResizability(.contentSize)
+        .commands {
+            CommandGroup(after: .appInfo) {
+                CheckForUpdatesView(updater: updateService.updater)
+            }
+        }
     }
 }